删除 修改

+++ club-auth/club-auth-infra/src/main/java/com/jinwang/auth/infra/basic/mapper/AuthUserDao.java
@@ -70,7 +70,7 @@ public interface AuthUserDao {
      * @param id 主键
      * @return 影响行数
      */
-    int deleteById(Integer id);
+    int deleteById(Long id);
 
     /**
      * 查询指定行数据
